add missing deps, merge .h and .cc of plugins
