+++ EventFilter/GEMRawToDigi/plugins/BuildFile.xml
@@ -1,8 +1,12 @@
 <use name="CondFormats/DataRecord"/>
 <use name="CondFormats/GEMObjects"/>
-<use name="EventFilter/GEMRawToDigi"/>
+<use name="DataFormats/Common"/>
 <use name="DataFormats/FEDRawData"/>
 <use name="DataFormats/GEMDigi"/>
+<use name="EventFilter/GEMRawToDigi"/>
+<use name="FWCore/Framework"/>
+<use name="FWCore/ParameterSet"/>
+<use name="FWCore/Utilities"/>
 <library file="*.cc" name="EventFilterGEMRawToDigiPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
