+++ EventFilter/GEMRawToDigi/plugins/GEMRawToDigiModule.cc
@@ -1,17 +1,57 @@
-/** \unpacker for gem
+/** \class GEMRawToDigiModule
+ *  \unpacker for gem
+ *  \based on CSCDigiToRawModule
  *  \author J. Lee - UoS
  */
+
+#include "CondFormats/DataRecord/interface/GEMeMapRcd.h"
+#include "CondFormats/GEMObjects/interface/GEMeMap.h"
+#include "CondFormats/GEMObjects/interface/GEMROMapping.h"
 #include "DataFormats/Common/interface/Handle.h"
+#include "DataFormats/FEDRawData/interface/FEDNumbering.h"
+#include "DataFormats/FEDRawData/interface/FEDRawDataCollection.h"
+#include "DataFormats/GEMDigi/interface/AMC13Event.h"
+#include "DataFormats/GEMDigi/interface/GEMAMC13EventCollection.h"
+#include "DataFormats/GEMDigi/interface/GEMAMCdataCollection.h"
+#include "DataFormats/GEMDigi/interface/GEMDigiCollection.h"
+#include "DataFormats/GEMDigi/interface/GEMGEBdataCollection.h"
+#include "DataFormats/GEMDigi/interface/GEMVfatStatusDigiCollection.h"
+#include "DataFormats/GEMDigi/interface/VFATdata.h"
+#include "EventFilter/GEMRawToDigi/interface/GEMRawToDigi.h"
 #include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/Run.h"
 #include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Run.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+#include "FWCore/Utilities/interface/ESGetToken.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 #include "FWCore/Utilities/interface/Transition.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "DataFormats/FEDRawData/interface/FEDNumbering.h"
 
-#include "EventFilter/GEMRawToDigi/plugins/GEMRawToDigiModule.h"
-#include "FWCore/Framework/interface/ESTransientHandle.h"
+class GEMRawToDigiModule : public edm::global::EDProducer<edm::RunCache<GEMROMapping> > {
+public:
+  /// Constructor
+  GEMRawToDigiModule(const edm::ParameterSet& pset);
+
+  // global::EDProducer
+  std::shared_ptr<GEMROMapping> globalBeginRun(edm::Run const&, edm::EventSetup const&) const override;
+  void produce(edm::StreamID, edm::Event&, edm::EventSetup const&) const override;
+  void globalEndRun(edm::Run const&, edm::EventSetup const&) const override{};
+
+  // Fill parameters descriptions
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  edm::EDGetTokenT<FEDRawDataCollection> fed_token;
+  edm::ESGetToken<GEMeMap, GEMeMapRcd> gemEMapToken_;
+  bool useDBEMap_;
+  bool unPackStatusDigis_;
+  std::unique_ptr<GEMRawToDigi> gemRawToDigi_;
+};
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(GEMRawToDigiModule);
 
 using namespace gem;
