--- EventFilter/GEMRawToDigi/plugins/GEMRawToDigiModule.h
@@ -1,51 +1,0 @@
-#ifndef EventFilter_GEMRawToDigi_GEMRawToDigiModule_h
-#define EventFilter_GEMRawToDigi_GEMRawToDigiModule_h
-
-/** \class GEMRawToDigiModule
- *  \based on CSCDigiToRawModule
- *  \author J. Lee - UoS
- */
-
-#include "FWCore/Framework/interface/global/EDProducer.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "FWCore/Utilities/interface/ESGetToken.h"
-#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
-#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
-
-#include "DataFormats/FEDRawData/interface/FEDRawDataCollection.h"
-#include "DataFormats/GEMDigi/interface/GEMDigiCollection.h"
-#include "DataFormats/GEMDigi/interface/GEMVfatStatusDigiCollection.h"
-#include "DataFormats/GEMDigi/interface/GEMGEBdataCollection.h"
-#include "DataFormats/GEMDigi/interface/GEMAMCdataCollection.h"
-#include "DataFormats/GEMDigi/interface/GEMAMC13EventCollection.h"
-#include "DataFormats/GEMDigi/interface/AMC13Event.h"
-#include "DataFormats/GEMDigi/interface/VFATdata.h"
-
-#include "EventFilter/GEMRawToDigi/interface/GEMRawToDigi.h"
-
-#include "CondFormats/DataRecord/interface/GEMeMapRcd.h"
-#include "CondFormats/GEMObjects/interface/GEMeMap.h"
-#include "CondFormats/GEMObjects/interface/GEMROMapping.h"
-
-class GEMRawToDigiModule : public edm::global::EDProducer<edm::RunCache<GEMROMapping> > {
-public:
-  /// Constructor
-  GEMRawToDigiModule(const edm::ParameterSet& pset);
-
-  // global::EDProducer
-  std::shared_ptr<GEMROMapping> globalBeginRun(edm::Run const&, edm::EventSetup const&) const override;
-  void produce(edm::StreamID, edm::Event&, edm::EventSetup const&) const override;
-  void globalEndRun(edm::Run const&, edm::EventSetup const&) const override{};
-
-  // Fill parameters descriptions
-  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
-
-private:
-  edm::EDGetTokenT<FEDRawDataCollection> fed_token;
-  edm::ESGetToken<GEMeMap, GEMeMapRcd> gemEMapToken_;
-  bool useDBEMap_;
-  bool unPackStatusDigis_;
-  std::unique_ptr<GEMRawToDigi> gemRawToDigi_;
-};
-DEFINE_FWK_MODULE(GEMRawToDigiModule);
-#endif
