+++ EventFilter/Phase2TrackerRawToDigi/plugins/BuildFile.xml
@@ -1,12 +1,14 @@
 <library file="*.cc" name="EventFilterPhase2TrackerRawToDigiPlugins">
-  <use name="CondFormats/SiStripObjects"/>
   <use name="CondFormats/DataRecord"/>
+  <use name="CondFormats/SiStripObjects"/>
   <use name="DataFormats/Common"/>
   <use name="DataFormats/DetId"/>
   <use name="DataFormats/FEDRawData"/>
+  <use name="DataFormats/Phase2TrackerDigi"/>
   <use name="EventFilter/Phase2TrackerRawToDigi"/>
   <use name="FWCore/Framework"/>
   <use name="FWCore/MessageLogger"/>
   <use name="FWCore/ParameterSet"/>
+  <use name="FWCore/Utilities"/>
   <flags EDM_PLUGIN="1"/>
 </library>
