--- EventFilter/Phase2TrackerRawToDigi/plugins/module.cc
@@ -1,10 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "EventFilter/Phase2TrackerRawToDigi/plugins/Phase2TrackerDigiProducer.h"
-#include "EventFilter/Phase2TrackerRawToDigi/plugins/Phase2TrackerCommissioningDigiProducer.h"
-
-typedef Phase2Tracker::Phase2TrackerDigiProducer Phase2TrackerDigiProducer;
-typedef Phase2Tracker::Phase2TrackerCommissioningDigiProducer Phase2TrackerCommissioningDigiProducer;
-
-DEFINE_FWK_MODULE(Phase2TrackerDigiProducer);
-DEFINE_FWK_MODULE(Phase2TrackerCommissioningDigiProducer);
